Set DAC to low-latency mode
To prevent an update of the offset DAC interfering with an update for the main
DAC during a single sample period, the main DAC is set to low-latency mode.
This means it will output data as soon as it arrives. Since the data is already
syncrhonised to the sample clock in the DSP module, this second syncronisation
is unnecessary.

This change eliminates phase jumps in the signal when the offset DAC is set.