Add dedicated command to wait for the filter reload to be ready
We need the filter reload signals to be high before resetting the
filters, else they will never come high. We also need to ensure that
the filters are ready for reload before we actually attempt a reload.
So it makes sense to have a dedicated command for waiting until the
filters are ready, which can be called before a reset or a reload.