Add scripts for generating filter coefficients and loading offsets
The DSP filters use windowed-sinc FIR filters. The simple low-pass
filter for calculating the voltage amplitude is generated and written
to the FPGA with the lpf_design.tcl script. The deconvolution filter
used to calculate real-time power is generated and written to the
FPGA by the power_filter_design.tcl script. Both scripts use the
firwin_kb.tcl script to actually design the filter kernel, before
converting to fixed point and formatting appropriately.

Additionally, the load_offset_channel.tcl script is used for
converting offset values in Volts (and sensitivity in V/W) for a
given channel into register values to be written to the correct
memory address to perform voltage and power offset correction
on-chip.

The program for actually calculating the offsets, sensitivity
and cooling time, which are needed for the power filter and
offset loading, are assumed to have already been calculated, either
online or offline. These scripts only work with the pre-calculated
values given as command line parameters.