Add utility to calculate calibration parameters on the device
The "calibfit" command will read transient capture data for the
last capture, which was assumed to be a calibration run consisting
of a waiting period, followed by continuous heating and then cooling
for the rest of the capture. It fits an exponential decay with
amplitude and offset to the cooling curve, and combines this with
the measured DC voltage and current to calculate the bolometer
calibration constants: sensitivity (V/W), cooling time and
voltage offsets.

It is rather simple to use because it assumes a fixed calibration
sequence and simple sensor physics, but should be enough to provide
fast calibration results (~0.3s per channel for 2.1s of calibration
data) to the user, for example to load the deconvolution filter and
offset registers before a shot.

Note that calibfit is a compiled C++ file, and uses the lmfit C
library for least-squares fitting
(http://apps.jcns.fz-juelich.de/doku/sc/lmfit). It thus requires
a cross-compiler: the Xilinx-provided gcc and g++ compilers are
assumed to be in the PATH (see the Makefile provided for details).