calibfit: improve extraction of cooling curves
The intention was always to extract all data from when cooling started
to the end of the trace. However, the previous implementation did not
technically do this: it merely extracted all elements below a certain
threshold after a given delay. The new implementation does exactly
what is asked: finds the first element below a threshold after a given
delay, assumes this is the start of the cooling, and copies all data
from there until the end of the trace as the cooling curve. This does
require the cooling threshold to be sufficiently low that noise during
the heating period (after t_wait) won't cause the heating signal to
drop below the threshold. In practice this won't be an issue.