Enable copying calibration data so it isn't overwritten
Calling wait_for_filters when the filter state is not 0x33 starts a
stream to nowhere to run some data through the filters. This is needed
to stop the system from locking up when trying to load more coefficients, or
when resetting the filters. However, this data stream overwrites the data
in `/dev/acq400/data/${site}/${channel}.`

This commit adds provision to copy the data to `/tmp/calib_data` before
loading the filters, so a copy of the data is still available until
run_calibfit is next run. The calibfit utility has also been updated
to be able to read from a user-specified directory (although the layout
should be the same as /dev/acq400/data), so that calibfit can be
run again after the filters have been loaded, for example with
different -H and -C flags.

To do this, add `COPY_CALIB_DATA=1` to `/mnt/local/sysconfig/bolo.sh`

	modified:   calibfit/calibfit.cxx
	modified:   usr/local/bin/run_calibfit