Fix output formatting of offsets RAM
Re-formatted the output to accurately reflect the contents of the
offsets RAM, which consists of I0/Q0 pairs followed by PI0/PQ0 pairs.

There's a bug in fs2xml which results in a double free if trying to
parse more than 43 channels, so output is limited to this for now.

Note that the DSP module on the FPGA is configured for 48 channels,
the maximum supported by an ACQ2106 (with 6x BOLO8). It may make more
sense to only display the RAM contents for channels in active sites,
but for now all channels are displayed (it's easy enough to ignore
ones not being used).

	modified:   usr/local/bin/diagnostics.tcl