Set software trigger source before running the calibration
The calibration routine uses soft triggering to both acquire the
calibration data and also to run a stream of data through the filters
when loading coefficients. The run_calibfit script needs to set up
the system to expect a software trigger.

Typically, systems will be configured to expect a hardware trigger
for the actual shot, and so the software trigger configuration needs
to be temporary. The triggering setup is restored once the calibration
routine is finished.