Remove mix of streaming and transient capture for running data through filters
If the acq400_streamd.0.conf file includes the line
`export ACQ400D_TRIG="soft_trigger"`, then data will start flowing
automatically after `streatonowhered start`. This removes the need
to manually check that the system is armed and then send a software
trigger (as long as `trg=1,1,1` is set).

This is not the default setup: despite it being set by the
default /mnt/local/rc.user, an end user could reasonably be expected
to modify rc.user so that it is no longer set. So provision is made
to configure automatic triggering for streaming data where required,
and then return the system to the original configuration afterwards.
This therefore places no restrictions on what should be in
/mnt/local/rc.user for the run_calibration and wait_for_filters scripts
to work properly.