Update power FIR coefficient reload order
Now that all ACQ2106 units have a 125 MHz FPGA clock, the power FIR
can be reconfigured for this clock for more efficient BRAM usage. But
this results in a change to the coefficient reload order. This commit
updates power_filter_design.tcl to match the updated reload order.

Note that this software package should only therefore be used with
newer bitstreams which have the power FIR filter configured in this
way.