Use transient instead of streaming in powerup
Now powerup runs after the transient package is loaded, we can use a
transient capture rather than start/stop a stream to set the DAC
going. This avoids the race condition of hoping the system arms itself
in the 1 second sleep before soft triggering the stream: a transient
can be configured to trigger as soon as it's armed.