Fix checking BOLO sysconfig variables in run_calibfit
Some tests were for whether the variable was set, rather than if it is
set to true (1) or false (0). In the case of BOLO_LOAD_OFFSETS, this
was actually set at the start of the script if it didn't exist,
providing no way to disable loading offsets.

This change sets all relevant sysconfig variables to a sensible
default value if they aren't set in /mnt/local/sysconfig/bolo.sh. For
boolean switches it then compares with the values 1 or 0, meaning
features like BOLO_VERBOSE, LOAD_POWER_FILTERS, BOLO_LOAD_OFFSETS and
COPY_CALIB_DATA can be explicitly enabled or disabled the sysconfig.